pyproject.toml
[project]
name = "fantasai-sports"
version = "0.1.0"
description = "Intelligent fantasy baseball assistant"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
    "sqlalchemy>=2.0",
    "alembic>=1.10",
    "psycopg2-binary>=2.9",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pybaseball>=2.2",
    "anthropic>=0.40",
    "httpx>=0.24",
    "python-dotenv>=1.0",
    "pandas>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
    "mypy>=1.0",
    "pandas-stubs>=2.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 99
target-version = "py39"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: live network tests — run with pytest -m integration",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = "pybaseball.*"
ignore_missing_imports = true
